scripts/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.8.30
cffi==1.17.1
exceptiongroup==1.2.2
iniconfig==2.0.0
minio==7.2.8
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pytest==8.3.3
pytest-mock==3.14.0
pytest-sugar==1.0.0
termcolor==2.4.0
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
